Underline the main menu buttons when hovering over them
Previously - like the index buttons - the main menu buttons would expand
when hovered over.

To improve consistency with the index buttons and improve the sleekness,
the buttons now preserve their dimensions, and instead are underlined
with a lines that match the font colour.

------------------------------------------------------------------------

Note: the margin & padding attributes had to be tweaked to preserve the
distances between buttons

Additionally, the borders were also modified to preserve the positioning
of the buttons between normal & hover states.