Enhance versatility & appearance of the projects table
This commit is similar in nature to 63224dd, in the sense of CSS and jQ
DataTables being used.

------------------------------------------------------------------------

Note: this commit removes the HTTP search form on the projects page to
favour the client-side search form. The back-end logic for searching the
projects is still present, although it's practically unused now.